Update config to not use hypeclair server anymore
